+++ src/nodes/features.cpp
@@ -36,6 +36,7 @@
 
 #include <cmath>
 #include "features.h"
+#include "trigger.h"
 
 /** @file
 
@@ -123,7 +124,9 @@ namespace
  */
 Features::Features(dc1394camera_t *camera):
   camera_(camera)
-{}
+{
+  trigger_.reset(new Trigger(camera));
+}
 
 /** Query and set all features for newly opened (or reopened) device.
  *
@@ -136,6 +139,8 @@ Features::Features(dc1394camera_t *camera):
  */
 bool Features::initialize(Config *newconfig)
 {
+  bool retval = true;
+
   // query all features for this device
   if (DC1394_SUCCESS != dc1394_feature_get_all(camera_, &feature_set_))
     {
@@ -166,14 +171,20 @@ bool Features::initialize(Config *newconfig)
             &newconfig->auto_sharpness, &newconfig->sharpness);
   configure(DC1394_FEATURE_SHUTTER,
             &newconfig->auto_shutter, &newconfig->shutter);
+  configure(DC1394_FEATURE_TRIGGER,
+            &newconfig->auto_trigger, &newconfig->trigger);
   configure(DC1394_FEATURE_WHITE_BALANCE, &newconfig->auto_white_balance,
             &newconfig->white_balance_BU, &newconfig->white_balance_RV);
   configure(DC1394_FEATURE_ZOOM,
             &newconfig->auto_zoom, &newconfig->zoom);
 
+  // set up trigger class, if supported by this camera
+  if (hasTrigger())
+    retval = trigger_->initialize(newconfig);
+
   // save configured values
   oldconfig_ = *newconfig;
-  return true;
+  return retval;
 }
 
 /** Reconfigure features for already open device.
@@ -223,6 +234,9 @@ void Features::reconfigure(Config *newconfig)
   updateIfChanged(DC1394_FEATURE_SHUTTER,
                   oldconfig_.auto_shutter, &newconfig->auto_shutter,
                   oldconfig_.shutter, &newconfig->shutter);
+  updateIfChanged(DC1394_FEATURE_TRIGGER,
+                  oldconfig_.auto_trigger, &newconfig->auto_trigger,
+                  oldconfig_.trigger, &newconfig->trigger);
   // White balance has two component parameters: Blue/U and Red/V.
   updateIfChanged(DC1394_FEATURE_WHITE_BALANCE,
                   oldconfig_.auto_white_balance,
@@ -232,6 +246,10 @@ void Features::reconfigure(Config *newconfig)
   updateIfChanged(DC1394_FEATURE_ZOOM,
                   oldconfig_.auto_zoom, &newconfig->auto_zoom,
  		  oldconfig_.zoom, &newconfig->zoom);
+
+  // reconfigure trigger class, if supported by this camera
+  if (hasTrigger())
+    trigger_->reconfigure(newconfig);
 
   // save modified values
   oldconfig_ = *newconfig;