+++ src/nodes/features.h
@@ -41,6 +41,7 @@
 #include <dc1394/dc1394.h>
 
 #include "camera1394/Camera1394Config.h"
+#include "trigger.h"
 typedef camera1394::Camera1394Config Config;
 
 /** @file
@@ -65,6 +66,11 @@ public:
   ~Features() {};
   bool initialize(Config *newconfig);
   void reconfigure(Config *newconfig);
+
+  inline bool isTriggerPowered()
+  {
+	  return trigger_->isPowered();
+  }
 
 private:
   typedef int state_t;      ///< camera1394::Camera1394_* state values
@@ -92,6 +98,20 @@ private:
       }
     return false;
   }
+
+  /** Does this camera support triggering?
+   *
+   *  @pre feature_set_ initialized for this camera
+   *  @return true if triggering supported
+   */
+  inline bool hasTrigger(void)
+  {
+    return DC1394_TRUE == feature_set_.feature[DC1394_FEATURE_TRIGGER
+                                               - DC1394_FEATURE_MIN].available;
+  }
+
+  // pointer to subordinate trigger class
+  boost::shared_ptr<Trigger> trigger_;
 
   bool setMode(dc1394feature_info_t *finfo, dc1394feature_mode_t mode);
   void setPower(dc1394feature_info_t *finfo, dc1394switch_t on_off);